Removing colors from cmdline options. Changing tests to not add cmdline options. changing formattedTask.writeString to use fmt.Fprintf. moved loadConfigCmd method into main.go, so config.go don't have a kingpin dependency.
