+++ .todo2
@@ -10,11 +10,6 @@
       "text": "Add support for syncing tasks with TODO entries in source.",
       "priority": "low",
       "creation": 1332544798
-    },
-    {
-      "text": "herpderp",
-      "priority": "veryhigh",
-      "creation": 1497563128
     }
   ]
 }
