+++ config_test.go
@@ -26,7 +26,7 @@ import (
 func TestLoadAllOptionsInCorrectOrderPrimarilyCMDLineOptions(t *testing.T) {
 	var fail bytes.Buffer
 	config := NewConfig()
-	os.Args = []string{"devtodo2", "-A", "--verylowfgcolor", BLACK, "--lowfgcolor", CYAN, "--mediumfgcolor", WHITE, "--highfgcolor", BRIGHTYELLOW, "--verylowbgcolor", BLACK, "--lowbgcolor", YELLOW, "--mediumbgcolor", WHITE, "--highbgcolor", BLUE, "--veryhighbgcolor", CYAN}
+	os.Args = []string{"devtodo2", "-A"}
 
 	loadConfigCMD(config)
 
@@ -44,7 +44,7 @@ func TestLoadAllOptionsInCorrectOrderPrimarilyCMDLineOptions(t *testing.T) {
 	}
 
 	veryLowFGColorFromConfig := config.FGColors[PriorityFromString(verylow)]
-	expectedVeryLowFGColorFromConfig := BLACK
+	expectedVeryLowFGColorFromConfig := BLUE
 	if veryLowFGColorFromConfig != expectedVeryLowFGColorFromConfig {
 		fail.WriteString(fmt.Sprintf("\n Err at config_test.go: config.FGColors.verylow Expected:%s , Got: %s \n", expectedVeryLowFGColorFromConfig, veryLowFGColorFromConfig))
 		t.Fail()
@@ -79,35 +79,35 @@ func TestLoadAllOptionsInCorrectOrderPrimarilyCMDLineOptions(t *testing.T) {
 	}
 
 	veryLowBGColorFromConfig := config.BGColors[PriorityFromString(verylow)]
-	expectedVeryLowBGColorFromConfig := BLACK
+	expectedVeryLowBGColorFromConfig := NOCOLOR
 	if veryLowBGColorFromConfig != expectedVeryLowBGColorFromConfig {
 		fail.WriteString(fmt.Sprintf("\n Err at config_test.go: config.BGColors.verylow  Expected:%s , Got: %s \n", expectedVeryLowBGColorFromConfig, veryLowBGColorFromConfig))
 		t.Fail()
 	}
 
 	lowBGColorFromConfig := config.BGColors[PriorityFromString(low)]
-	expectedLowBGColorFromConfig := YELLOW
+	expectedLowBGColorFromConfig := NOCOLOR
 	if lowBGColorFromConfig != expectedLowBGColorFromConfig {
 		fail.WriteString(fmt.Sprintf("\n Err at config_test.go: config.BGColors.low  Expected:%s , Got: %s \n", expectedLowBGColorFromConfig, lowBGColorFromConfig))
 		t.Fail()
 	}
 
 	mediumBGColorFromConfig := config.BGColors[PriorityFromString(medium)]
-	expectedMediumBGColorFromConfig := WHITE
+	expectedMediumBGColorFromConfig := NOCOLOR
 	if mediumBGColorFromConfig != expectedMediumBGColorFromConfig {
 		fail.WriteString(fmt.Sprintf("\n Err at config_test.go: config.BGColors.medium  Expected:%s , Got: %s \n", expectedMediumBGColorFromConfig, mediumBGColorFromConfig))
 		t.Fail()
 	}
 
 	highBGColorFromConfig := config.BGColors[PriorityFromString(high)]
-	expectedHighBGColorFromConfig := BLUE
+	expectedHighBGColorFromConfig := NOCOLOR
 	if highBGColorFromConfig != expectedHighBGColorFromConfig {
 		fail.WriteString(fmt.Sprintf("\n Err at config_test.go: config.BGColors.high  Expected:%s , Got: %s \n", expectedHighBGColorFromConfig, highBGColorFromConfig))
 		t.Fail()
 	}
 
 	veryHighBGColorFromConfig := config.BGColors[PriorityFromString(veryhigh)]
-	expectedVeryHighBGColorFromConfig := CYAN
+	expectedVeryHighBGColorFromConfig := NOCOLOR
 	if veryHighBGColorFromConfig != expectedVeryHighBGColorFromConfig {
 		fail.WriteString(fmt.Sprintf("\n Err at config_test.go: config.BGColors.veryhigh  Expected:%s , Got: %s \n", expectedVeryHighBGColorFromConfig, veryHighBGColorFromConfig))
 		t.Fail()
@@ -121,7 +121,7 @@ func TestLoadAllOptionsInCorrectOrderPrimarilyConfigurationFileOptions(t *testin
 	config := NewConfig()
 
 	priority := veryhigh
-	os.Args = []string{"devtodo2", "-p", priority, "--verylowbgcolor", RED, "--highfgcolor", BRIGHTCYAN, "-a", "Test task name whatever."}
+	os.Args = []string{"devtodo2", "-p", priority, "-a", "Test task name whatever."}
 
 	loadConfigCMD(config)
 
@@ -160,7 +160,7 @@ func TestLoadAllOptionsInCorrectOrderPrimarilyConfigurationFileOptions(t *testin
 	}
 
 	highFGColorFromConfig := config.FGColors[PriorityFromString(high)]
-	expectedHighFGColorFromConfig := BRIGHTCYAN
+	expectedHighFGColorFromConfig := BRIGHTYELLOW
 	if highFGColorFromConfig != expectedHighFGColorFromConfig {
 		fail.WriteString(fmt.Sprintf("\n Err at config_test.go: config.FGColors.high  Expected:%s , Got: %s \n", expectedHighFGColorFromConfig, highFGColorFromConfig))
 		t.Fail()
@@ -174,7 +174,7 @@ func TestLoadAllOptionsInCorrectOrderPrimarilyConfigurationFileOptions(t *testin
 	}
 
 	veryLowBGColorFromConfig := config.BGColors[PriorityFromString(verylow)]
-	expectedVeryLowBGColorFromConfig := RED
+	expectedVeryLowBGColorFromConfig := NOCOLOR
 	if veryLowBGColorFromConfig != expectedVeryLowBGColorFromConfig {
 		fail.WriteString(fmt.Sprintf("\n Err at config_test.go: config.BGColors.verylow  Expected:%s , Got: %s \n", expectedVeryLowBGColorFromConfig, veryLowBGColorFromConfig))
 		t.Fail()
